network_to_diagram/requirements.txt
netmiko>=4.2.0
networkx>=3.1
matplotlib>=3.7.1
pydot>=1.4.2
graphviz>=0.20.1
pysnmp>=4.4.12
scapy>=2.5.0
pyyaml>=6.0
